# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DeadlineRow": {
        "properties": {
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "classification": {
            "$ref": "#/components/schemas/MilestoneClass"
          },
          "days_until_due": {
            "title": "Days Until Due",
            "type": "integer"
          },
          "due_date": {
            "format": "date",
            "title": "Due Date",
            "type": "string"
          },
          "listing_key": {
            "title": "Listing Key",
            "type": "string"
          },
          "milestone_id": {
            "title": "Milestone Id",
            "type": "string"
          },
          "milestone_type": {
            "$ref": "#/components/schemas/MilestoneType"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "transaction_id",
          "milestone_id",
          "milestone_type",
          "title",
          "due_date",
          "classification",
          "days_until_due",
          "listing_key"
        ],
        "title": "DeadlineRow",
        "type": "object"
      },
...
      "Document": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file": {
            "$ref": "#/components/schemas/FileRef"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/DocumentKind",
            "default": "other"
          },
          "milestone_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Milestone Id"
          },
          "tenant_id": {
            "default": "",
            "title": "Tenant Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          },
          "uploaded_by": {
            "default": "",
            "title": "Uploaded By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "transaction_id",
          "title",
          "file",
          "created_at"
        ],
        "title": "Document",
        "type": "object"
      },
      "DocumentKind": {
        "enum": [
          "purchase_agreement",
          "disclosures",
          "inspection_report",
          "appraisal_report",
          "loan_approval",
          "closing_statement",
          "other"
        ],
        "title": "DocumentKind",
        "type": "string"
      },
...
      "FileRef": {
        "description": "A pointer to a file living in the office file store.\n\n`file_id` is the backend's opaque identifier; `web_url` is where a human\nopens the file (the provider's own viewer). No provider-specific fields.",
        "properties": {
          "file_id": {
            "title": "File Id",
            "type": "string"
          },
          "mime_type": {
            "default": "",
            "title": "Mime Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "web_url": {
            "default": "",
            "title": "Web Url",
            "type": "string"
          }
        },
        "required": [
          "file_id",
          "name"
        ],
        "title": "FileRef",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MilestoneClass": {
        "enum": [
          "on_track",
          "due_soon",
          "overdue",
          "blocked_external"
        ],
        "title": "MilestoneClass",
        "type": "string"
      },
      "MilestoneStatus": {
        "enum": [
          "pending",
          "complete",
          "waived"
        ],
        "title": "MilestoneStatus",
        "type": "string"
      },
      "MilestoneType": {
        "enum": [
          "inspection",
          "appraisal",
          "financing",
          "closing",
          "custom"
        ],
        "title": "MilestoneType",
        "type": "string"
      },
      "MilestoneView": {
        "properties": {
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "days_until_due": {
            "title": "Days Until Due",
            "type": "integer"
          },
          "document_satisfied": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Satisfied"
          },
          "due_date": {
            "format": "date",
            "title": "Due Date",
            "type": "string"
          },
          "escalation_level": {
            "default": 0,
            "title": "Escalation Level",
            "type": "integer"
          },
          "expected_document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "owner": {
            "default": "",
            "title": "Owner",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MilestoneStatus",
            "default": "pending"
          },
          "tenant_id": {
            "default": "",
            "title": "Tenant Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/MilestoneType"
          }
        },
        "required": [
          "id",
          "transaction_id",
          "type",
          "title",
          "due_date",
          "classification",
          "days_until_due"
        ],
        "title": "MilestoneView",
        "type": "object"
      },
...
      "OpenTransaction": {
        "properties": {
          "close_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Date"
          },
          "contract_date": {
            "format": "date",
            "title": "Contract Date",
            "type": "string"
          },
          "listing_key": {
            "title": "Listing Key",
            "type": "string"
          },
          "parties": {
            "items": {
              "$ref": "#/components/schemas/TransactionParty"
            },
            "title": "Parties",
            "type": "array"
          }
        },
        "required": [
          "listing_key",
          "contract_date"
        ],
        "title": "OpenTransaction",
        "type": "object"
      },
...
      "Transaction": {
        "properties": {
          "close_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Date"
          },
          "contract_date": {
            "format": "date",
            "title": "Contract Date",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "listing_key": {
            "title": "Listing Key",
            "type": "string"
          },
          "parties": {
            "items": {
              "$ref": "#/components/schemas/TransactionParty"
            },
            "title": "Parties",
            "type": "array"
          },
          "stage": {
            "$ref": "#/components/schemas/TransactionStage"
          },
          "tenant_id": {
            "default": "",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "listing_key",
          "stage",
          "contract_date"
        ],
        "title": "Transaction",
        "type": "object"
      },
      "TransactionDetail": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "title": "Documents",
            "type": "array"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneView"
            },
            "title": "Milestones",
            "type": "array"
          },
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          }
        },
        "required": [
          "transaction",
          "milestones",
          "documents"
        ],
        "title": "TransactionDetail",
        "type": "object"
      },
      "TransactionParty": {
        "properties": {
          "contact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "email": {
            "default": "",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "name"
        ],
        "title": "TransactionParty",
        "type": "object"
      },
      "TransactionSearchRow": {
        "properties": {
          "property_address": {
            "title": "Property Address",
            "type": "string"
          },
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          }
        },
        "required": [
          "transaction",
          "property_address"
        ],
        "title": "TransactionSearchRow",
        "type": "object"
      },
      "TransactionStage": {
        "enum": [
          "under_contract",
          "contingencies",
          "clear_to_close",
          "closed",
          "cancelled"
        ],
        "title": "TransactionStage",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "brokerops api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/transactions": {
      "get": {
        "operationId": "list_transactions_transactions_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionDetail"
                  },
                  "title": "Response List Transactions Transactions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Transactions",
        "tags": [
          "transactions"
        ]
      },
      "post": {
        "description": "Open an escrow for a listing and generate its milestone timeline.\n\nIdempotent per listing: the id is derived from the listing key, so a repeat\nwith the same terms returns the existing transaction (200) instead of opening\na second (201); a repeat with different terms is a 409. Invalid escrow dates\nare rejected (422) before anything is persisted. V1 assumes one transaction\nper listing.",
        "operationId": "open_transaction_transactions_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTransaction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/deadlines": {
      "get": {
        "description": "The coordinator's cross-transaction deadline queue (BOP-030).\n\nEvery active transaction's pending milestones, classified due-soon / overdue /\nblocked-external and sorted most-urgent-first. The classification and ordering\nrules live in the core milestone engine; this route only enumerates and joins\nthe listing key for display.",
        "operationId": "deadline_queue_transactions_deadlines_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeadlineRow"
                  },
                  "title": "Response Deadline Queue Transactions Deadlines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deadline Queue",
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/search": {
      "get": {
        "description": "Find active transactions by listing key, party (contact) name, or property\naddress (BOP-030 viewer home).\n\nA thin case-insensitive substring match over the active transactions \u2014 the same\nworking set the board and the deadline queue read. A blank query returns\nnothing. The response is filtered to the caller's role (BOP-027), so a viewer\nsees party names but not their contact emails.",
        "operationId": "search_transactions_transactions_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionSearchRow"
                  },
                  "title": "Response Search Transactions Transactions Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Transactions",
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/{transaction_id}": {
      "get": {
        "operationId": "get_transaction_transactions__transaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
...
  }
}